// interfaces/Book.ts
export type Book = {
  authors: string;
  code1: string;
  code2: string;
  description: string;
  id: string;
  infolink: string;
  language: string;
  maturityRating: string;
  pageCount: number;
  previewLink: string;
  printType: string;
  publishedDate: string;
  publisher: string;
  smallThumbnail: string;
  thumbnail: string;
  title: string;
  canonicalLink: string;
  onShelf: string;
  subtitle: string;
  categories: string;
  averageRating: number;
  ratingsCount: number;
}